// src/types/postList.ts
export type postProps = {
  _id: string;
  id?: number;
  image?: string;
  title?: string;
  date?: string;
  desc: string;
  category?: string;
  heading1?: string;
  paragraph1?: string;
  heading2?: string;
  paragraph2?: string;
  heading3?: string;
  paragraph3?: string;
  heading4?: string;
  paragraph4?: string;
  quotes?: string;
  likes?: string;
};


// export type PostProps = {
//   id: string; // Primary identifier
//   legacyId?: number; // Secondary or legacy identifier
//   imageUrl?: string; // URL of the image associated with the post
//   title?: string; // Title of the post
//   date?: string; // Date when the post was published
//   desc: string; // Description or summary of the post
//   category?: string; // Category or tag associated with the post
//   introTitle?: string; // Title of the introduction section
//   introText?: string; // Content of the introduction section
//   mainTitle?: string; // Title of the main section
//   mainText?: string; // Content of the main section
//   subTitle?: string; // Title of the secondary section
//   subText?: string; // Content of the secondary section
//   conclTitle?: string; // Title of the conclusion section
//   conclText?: string; // Content of the conclusion section
//   quotes?: string; // Notable quotes from the post
//   likes?: string; // Number of likes the post has received
// };